EL-3430 - Timelne Chart Build Fix
