+++ src/plugins/index.ts
@@ -1,3 +1,2 @@
-export * from './chartjs/timeline/timeline-chart.interface';
 export * from './chartjs/timeline/timeline-chart.module';
